Flytt all kode som har med endringsresultat ut i egen fil (også for tester)
